lobby_request+invit
unfinished
idea : only inv creator lock lobby_request

+++ _bin/sql/func/lobby.sql
@@ -51,15 +51,30 @@ BEGIN
 END
 $$ LANGUAGE plpgsql;
 
+/*
+0 - request created / updated
+1 - lobby joined
+*/
 CREATE OR REPLACE FUNCTION join_lobby(_id_viewer integer, _id_lobby integer) RETURNS integer AS $$
+DECLARE
+  __filter_join bool;
 BEGIN
   PERFORM lock_lobbyuser(false, _id_viewer);
+  SELECT filter_join INTO __filter_join FROM lobbys WHERE id = _id_lobby FOR SHARE; PERFORM raise_except(NOT FOUND, 'lobby not found');
+  IF __filter_join THEN
+    PERFORM FROM lobby_members WHERE id_user=_id_viewer AND id_lobby=_id_lobby; PERFORM raise_except(FOUND, 'already lobby member');
+    DELETE FROM lobby_requests WHERE id_lobby = _id_lobby AND id_user = _id_viewer AND status = 'wait_user';
+    IF NOT FOUND THEN
+      INSERT INTO lobby_requests (id_lobby, id_user, status) VALUES (_id_lobby, _id_viewer, 'wait_lobby')
+        ON CONFLICT DO UPDATE SET id_creator = NULL WHERE id_creator IS NOT NULL;
+      PERFORM raise_except(NOT FOUND, 'request already exist');
+      RETURN 0;
+    END IF;
+  END IF;
   
-  PERFORM FROM lobbys WHERE id = _id_lobby FOR SHARE; PERFORM raise_except(NOT FOUND, 'lobby not found');
   UPDATE lobby_slots SET free_slots=free_slots-1 WHERE id_lobby = _id_lobby;
   INSERT INTO lobby_members(id_lobby, id_user) VALUES(_id_lobby, _id_viewer);
-  
-  DELETE FROM lobby_invitations WHERE id_lobby = _id_lobby AND id_target = _id_viewer;
+  DELETE FROM lobby_requests WHERE id_lobby = _id_lobby AND id_user = _id_viewer; -- CASCADE DELETE FROM lobby_invitations WHERE id_lobby = _id_lobby AND id_target = _id_viewer;
   RETURN 1;
 END
 $$ LANGUAGE plpgsql;
@@ -93,7 +108,7 @@ BEGIN
         SELECT id_user FROM lobby_members WHERE id_lobby = __id_lobby LIMIT 1
       ) SELECT lobby_members.id_user INTO __id_owner
                                      FROM lobby_members, cte_new_owner
-                                     WHERE id_lobby = __id_lobby AND lobby_members.id_user = cte_new_owner.id_user FOR KEY SHARE; -- serialization error can occur
+                                     WHERE id_lobby = __id_lobby AND lobby_members.id_user = cte_new_owner.id_user FOR KEY SHARE;
       UPDATE lobbys SET id_owner = __id_owner WHERE id = __id_lobby;
       RETURN 3;
     END IF;
@@ -110,14 +125,76 @@ BEGIN
   PERFORM lock_lobbyuser(true, _id_target);
   SELECT (id_owner = _id_viewer OR filter_join IS FALSE) INTO __authz FROM lobbys WHERE id = _id_lobby FOR SHARE; PERFORM raise_except(NOT __authz, 'lobby authz fail');
   PERFORM FROM lobby_members WHERE id_lobby = _id_lobby AND id_user = _id_target; PERFORM raise_except(FOUND, 'target already member');
+  
+  INSERT INTO lobby_requests(id_lobby, id_user, status, id_creator)
+    VALUES(_id_lobby, _id_target, CASE WHEN __authz THEN 'wait_user' ELSE 'wait_lobby' END, _id_viewer)
+    ON CONFLICT DO UPDATE SET status=CASE status WHEN 'wait_user' THEN 'wait_user' ELSE EXCLUDED.status END;
   INSERT INTO lobby_invitations(id_lobby, id_target, id_creator) VALUES (_id_lobby, _id_target, _id_viewer);
   RETURN FOUND;
 END
 $$ LANGUAGE plpgsql;
 
+--invit creator is the only one locking with lobby_request
 CREATE OR REPLACE FUNCTION lobby_invite_cancel(_id_viewer integer, _id_target integer) RETURNS boolean AS $$
+DECLARE
+  __id_lobby integer;
+  __id_creator integer;
 BEGIN
-  DELETE FROM lobby_invitations WHERE id_creator = _id_viewer AND id_target = _id_target;
+  SELECT id_lobby INTO __id_lobby FROM lobby_members WHERE id_user = _id_viewer;
+  
+  DELETE FROM lobby_invitations WHERE id_lobby = __id_lobby AND id_target = _id_target AND id_creator = _id_viewer; PERFORM raise_except(NOT FOUND, 'invit not found');
+  PERFORM FROM lobby_requests WHERE id_lobby=__id_lobby AND id_user = _id_target AND id_creator = _id_viewer FOR SHARE;
+  IF FOUND THEN
+    SELECT id_target INTO __id_creator FROM lobby_invitations WHERE id_lobby = __id_lobby AND id_target = _id_target LIMIT 1 FOR KEY SHARE SKIP LOCKED ; --for key share?
+    IF FOUND THEN --still invit left
+      UPDATE lobby_requests SET id_creator = __id_creator WHERE id_lobby = __id_lobby AND id_user = _id_target;
+    ELSE --no lobby invit left
+      DELETE FROM lobby_requests WHERE id_lobby = __id_lobby AND id_user = _id_target;
+    END IF;
+  END IF;
   RETURN FOUND;
 END
 $$ LANGUAGE plpgsql;
+
+/*
+CREATE OR REPLACE FUNCTION lobby_invite_cancel(_id_viewer integer, _id_target integer) RETURNS boolean AS $$
+DECLARE
+  __id_lobby integer;
+  __id_creator integer;
+BEGIN
+  SELECT id_lobby INTO __id_lobby FROM lobby_members WHERE id_user = _id_viewer;
+  SELECT id_creator = _id_viewer INTO __id_creator FROM lobby_requests WHERE id_lobby=__id_lobby AND id_user = _id_target FOR SHARE;
+  PERFORM FROM lobby_requests WHERE id_lobby=__id_lobby AND id_user = _id_target AND id_creator = _id_viewer FOR UPDATE;
+  DELETE FROM lobby_invitations WHERE id_lobby = __id_lobby AND id_creator = _id_viewer AND id_target = _id_target; PERFORM raise_except(NOT FOUND, 'invit not found');
+  IF __id_creator = _id_viewer THEN
+    SELECT id_target INTO __id_creator FROM lobby_invitations WHERE id_lobby = __id_lobby AND id_target = _id_target LIMIT 1;
+    IF FOUND THEN --still invit left
+      UPDATE lobby_requests SET id_creator = __id_creator WHERE id_lobby = __id_lobby AND id_user = _id_target;
+    ELSE --no lobby invit left
+      DELETE FROM lobby_requests WHERE id_lobby = __id_lobby AND id_user = _id_target;
+    END IF;
+  END IF;
+  RETURN FOUND;
+END
+$$ LANGUAGE plpgsql;
+*/
+
+DROP FUNCTION IF EXISTS lobbyuser_del_invit;
+CREATE OR REPLACE FUNCTION lobbyuser_del_invit(_id_lobby integer, _id_user integer) RETURNS void AS $$
+DECLARE
+  __new_creator integer;
+  __tmp_loop integer;
+BEGIN
+  FOR __tmp_loop IN DELETE FROM lobby_invitations WHERE id_lobby = _id_lobby AND id_creator = _id_user RETURNING id_target
+  LOOP
+    PERFORM FROM lobby_requests WHERE id_lobby = _id_lobby AND id_user = __tmp_loop AND id_creator = _id_user FOR SHARE;
+    SELECT id_creator INTO __new_creator FROM lobby_invitations WHERE id_lobby = _id_lobby AND id_target = __tmp_loop LIMIT 1 FOR KEY SHARE SKIP LOCKED;
+    IF FOUND THEN
+      UPDATE lobby_requests SET id_creator = __new_creator  WHERE id_lobby = _id_lobby AND id_user = __tmp_loop ;
+    ELSE
+      DELETE FROM lobby_requests WHERE id_lobby = _id_lobby AND id_user = __tmp_loop;
+    END IF;
+  END LOOP;
+END
+$$ LANGUAGE plpgsql;
+
